// src/types/enums/meeting.enum.ts
export enum MeetingCategory {
  PROJECT_KICKOFF = "Project Kickoff",
  STATUS_UPDATE = "Status Update",
  PLANNING = "Planning",
  PROBLEM_SOLVING = "Problem-Solving",
  DECISION_MAKING = "Decision-Making",
  PERFORMANCE_REVIEW = "Performance Review",
  STRATEGIC = "Strategic",
  REGULAR_TEAM = "Regular Team",
  CLIENT = "Client",
  TRAINING = "Training",
  OTHER = "Other",
}